+++ API/src/main/java/fr/maxlego08/essentials/api/enchantment/Enchantments.java
@@ -1,0 +1,14 @@
+package fr.maxlego08.essentials.api.enchantment;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface Enchantments {
+
+    Optional<EssentialsEnchantment> getEnchantments(String enchantment);
+
+    void register();
+
+    List<String> getEnchantments();
+
+}