+++ API/src/main/java/fr/maxlego08/essentials/api/messages/Message.java
@@ -264,6 +264,7 @@ public enum Message {
     DESCRIPTION_VOTE_SET("Set the number of votes for a player"),
     DESCRIPTION_VOTE_ADD("Add votes to a player"),
     DESCRIPTION_VOTE_REMOVE("Remove votes from a player"),
+    DESCRIPTION_ENCHANT("Add enchant to an item"),
 
 
     YOU("you"),
@@ -666,6 +667,15 @@ public enum Message {
     COMMAND_VAULT_RENAME_SUCCESS("<success>You just set the vault name to &f%name%<success>."),
     COMMAND_VAULT_RENAME_ERROR("<error>You cannot use this name for your vault."),
     COMMAND_VAULT_RENAME_RESET("<success>You just reset the name of your vault."),
+
+    COMMAND_ENCHANT_INVALID("<error>You need to specify a valid player."),
+    COMMAND_ENCHANT_ERROR_ENCHANT("<error>Impossible to find the enchantment &f%name%<error>."),
+    COMMAND_ENCHANT_ERROR_ITEM_SELF("<error>You must have an item in your hand."),
+    COMMAND_ENCHANT_ERROR_ITEM_PLAYER("<error>%player% has no item in his hand."),
+    COMMAND_ENCHANT_REMOVE_SELF("<success>The enchantment &f%enchant% <success>has been removed from your item in hand."),
+    COMMAND_ENCHANT_REMOVE_PLAYER("<success>The enchantment &f%enchant% <success>has been removed from the item from &f%player%’s<succes> hand."),
+    COMMAND_ENCHANT_SUCCESS_SELF("<success>The enchantment &f%enchant% <success>has been applied to your item in hand."),
+    COMMAND_ENCHANT_SUCCESS_PLAYER("<success>The enchantment &f%enchant% <success>has been applied to the item from &f%player%’s<succes> hand."),
 
     ;