+++ src/main/java/fr/maxlego08/essentials/commands/CommandLoader.java
@@ -79,6 +79,7 @@ import fr.maxlego08.essentials.commands.commands.utils.CommandRepairAll;
 import fr.maxlego08.essentials.commands.commands.utils.CommandRules;
 import fr.maxlego08.essentials.commands.commands.utils.CommandTrash;
 import fr.maxlego08.essentials.commands.commands.utils.CommandVersion;
+import fr.maxlego08.essentials.commands.commands.utils.admins.CommandEnchant;
 import fr.maxlego08.essentials.commands.commands.utils.admins.CommandFly;
 import fr.maxlego08.essentials.commands.commands.utils.admins.CommandGod;
 import fr.maxlego08.essentials.commands.commands.utils.admins.CommandInvsee;
@@ -191,7 +192,7 @@ public class CommandLoader {
 
         register("sethome", CommandSetHome.class, "hcreate", "hc");
         register("delhome", CommandDelHome.class, "hdelete", "hd");
-        register("home", CommandHome.class, "h");
+        register("home", CommandHome.class, "h", "homes");
 
         register("ban", CommandBan.class);
         register("mute", CommandMute.class);
@@ -223,6 +224,7 @@ public class CommandLoader {
         register("killall", CommandKillAll.class);
         register("seen", CommandSeen.class, "whois");
         register("seenip", CommandSeenIp.class, "whoisip");
+        register("enchant", CommandEnchant.class, "enchantment");
 
         register("kit", CommandKit.class, "kits");
         register("showkit", CommandShowKit.class);