+++ src/main/java/fr/maxlego08/essentials/commands/commands/essentials/CommandEssentialsConvert.java
@@ -4,6 +4,7 @@ import fr.maxlego08.essentials.api.EssentialsPlugin;
 import fr.maxlego08.essentials.api.commands.CommandResultType;
 import fr.maxlego08.essentials.api.commands.Permission;
 import fr.maxlego08.essentials.convert.cmi.CMIConvert;
+import fr.maxlego08.essentials.convert.essentialsx.EssentialsXConvert;
 import fr.maxlego08.essentials.zutils.utils.commands.VCommand;
 
 import java.util.Arrays;
@@ -25,6 +26,14 @@ public class CommandEssentialsConvert extends VCommand {
 
             CMIConvert cmiConvert = new CMIConvert(plugin);
             cmiConvert.convert(sender);
+
+            return CommandResultType.SUCCESS;
+        }
+
+        if (pluginName.equalsIgnoreCase("EssentialsX")) {
+
+            EssentialsXConvert essentialsXConvert = new EssentialsXConvert(plugin);
+            essentialsXConvert.convert(sender);
 
             return CommandResultType.SUCCESS;
         }