+++ changelog.md
@@ -1,3 +1,15 @@
+# Unreleased
+
+# 1.0.0.2
+
+- Added the command `/ess convert EssentialsX`, allows converting the database from EssentialsX to zEssentials.
+  Documentation: https://zessentials.groupez.dev/getting-started/convert#essentialsx
+- Fixed the scoreboard title not appearing
+- Added `/homes` alias for `/home` command for default configurations
+- Added `/enchant` command
+- Added enchantments list with aliases
+- Added dutch translation [#19](https://github.com/Maxlego08/zEssentials/pull/19)
+
 # 1.0.0.1
 
 - Added the command `/ess convert CMI`, allows converting the database from CMI to zEssentials.